+++ src/DpressServices.php
@@ -39,6 +39,7 @@ use Dynart\Dpress\Content\LinkTargetResolverInterface;
 use Dynart\Dpress\Content\LinkTargets;
 use Dynart\Dpress\Content\MarkdownRenderer;
 use Dynart\Dpress\Content\Shortcode\BreakShortcode;
+use Dynart\Dpress\Content\Shortcode\AudioShortcode;
 use Dynart\Dpress\Content\Shortcode\VideoShortcode;
 use Dynart\Dpress\Content\ShortcodeRenderer;
 use Dynart\Dpress\Block\Blocks;
@@ -217,6 +218,7 @@ class DpressServices {
         Micro::add(CodeBlockRenderer::class);
         Micro::add(ShortcodeRenderer::class);
         Micro::add(VideoShortcode::class);
+        Micro::add(AudioShortcode::class);
         Micro::add(BreakShortcode::class);
         Micro::add(Slugger::class);
         Micro::add(TreeOrder::class);
@@ -316,6 +318,7 @@ class DpressServices {
      */
     const SHORTCODES = [
         'video' => [[VideoShortcode::class, 'render'], Shortcodes::BLOCK],
+        'audio' => [[AudioShortcode::class, 'render'], Shortcodes::BLOCK],
         'br'    => [[BreakShortcode::class, 'render'], Shortcodes::INLINE],
     ];